requirements.txt
python-telegram-bot==21.9
python-dotenv==1.0.0
openai==1.55.3
PyMuPDF
# gtts==2.5.4
# flask==3.0.0